fix(stella-autonomy): harden the decision core against absurd and adversarial input
Four small correctness/coverage gaps found by a full-crate review of the
self-driving decision core:

- `priority::rank_of` narrowed a ladder's Vec position through
  `u8::try_from`, so an operator's ladder past 256 rungs turned "the
  ladder knows this label" into `None` — read by every caller as
  "nobody judged this yet", the exact confusion this module exists to
  remove. Widened the return type to `usize`, which matches what
  `Iterator::position` already returns and cannot fail to represent any
  ladder a `Vec` can hold.
- `PrState::CiPending`'s doc comment now says plainly that it also
  covers "CI came back green but Mergeability is still Unknown" — the
  action (`Wait`) was already correct, only the name read as narrower
  than the state actually is.
- `SessionStats::closures_balance` and `filings_balance` summed three
  `u32` counters read straight off a session's stats file with no
  range check; three adversarial or corrupted counters near `u32::MAX`
  overflowed the sum (a panic in debug, a silent wraparound in
  release). Promoted both sums to `u64`.
- Closed four test gaps the same review named: `EscalationReason::
  RebaseCeilingReached` is now driven through `deliver_next` directly,
  not only a serde round trip; `liveness` has a boundary test at
  `now - heartbeat == stale_after` and one second past it;
  `metrics(&[])` is tested on an empty ledger; `plan_cycle`'s heavy
  tier is tested at its exact `mem_total_gb == 32` /
  `disk_free_gb == 100` floors.

Witness mechanism: the `rank_of` and stats-overflow fixes were each
verified against the pre-fix production code with the new test kept in
place — `rank_of` fails to compile (`literal out of range for u8`) and
the stats sums panic with `attempt to add with overflow`, both restored
to green by the fix. The `CiPending` doc/state and the four coverage
tests pin behaviour the machine already had correct; they close a gap
in what proves it, per the issue's own item 4.

`stella-autonomy` stays a leaf crate — no new dependency.

Closes #5319

+++ crates/stella-cli/src/self_driving_cmd/state.rs
@@ -288,7 +288,7 @@ impl LoopState {
             .iter()
             .map(|issue| {
                 let rank = stella_autonomy::priority::rank_of(issue, ladder)
-                    .and_then(|i| ladder.rungs.get(usize::from(i)))
+                    .and_then(|i| ladder.rungs.get(i))
                     .cloned()
                     .unwrap_or_else(|| "untriaged".to_owned());
                 json!({
